requirements.txt
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
crayons==0.4.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.5.4
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
notebook==6.4.8
numpy==1.19.5
opencv-contrib-python==4.5.2.54
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.11.2
pylint==2.6.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
Stegano==0.9.7
streamlit==1.5.1
terminado==0.12.1
testpath==0.5.0
tifffile==2020.9.3
toml==0.10.1
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.1
typing_extensions==4.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
